+++ tests_integration/test_data/projects/napp_templates_processors_v2/app/another_script.sql
@@ -1,4 +1,4 @@
--- This file uses old templating syntax
+-- This file uses old templates syntax
 CREATE OR REPLACE TABLE &{ ctx.env.schema_name }.&{ ctx.env.table_name } (
     name STRING
 );